[FIX] account_invoice_extract,hr_expense_extract,iap_extract: Fix OCR note author
Before this commit, when the content of the expense is updated by OCR, the author might not be "OdooBot".
This happened because while filling the document with results (in hr_expense._predict_product)
there is a savepoint that removes the author.
This commit sets the OCR note author after filling the document with results,
which makes sure that the correct author (i.e., OdooBot) is always set.

Moreover, if the expense receipt has a dot in its name, the _predict_product was never visited.
So, the attachment name extraction is fixed to make sure that only file extension was removed.

This commit also adds unit tests to expense and invoice OCR extract to make sure that their notes are authored by OdooBot.
The test was not introduced in hr.applicant, however, since OCR does not change any tracked fields (i.e., OCR never logs a note on the chatter).

task-4197413

closes odoo/enterprise#76225

X-original-commit: c4a0a4a06b46ba4173024cb07f4666d466306bb9
Signed-off-by: Louis Baudoux (lba) <lba@odoo.com>
Signed-off-by: Hesham Saleh (hsal) <hsal@odoo.com>