[IMP] hr_recruitment_sign: improve test tour for recruitment
This PR improves the test tour for the recruitment module. As per the task
requirements, the default group by in all application menus has been removed,
which caused the test tour to fail. To address this, the step that opens the
grouped list has been removed from the tour to ensure it passes successfully.

task-4222504

closes odoo/enterprise#75798

Related: odoo/odoo#182877
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>