[IMP] l10n_eu_oss_reports: adjust tests for new rates
Description of the issue this commit addresses:

The community commit linked to this one changed the rates of luxembourgian
taxes used in the oss fiscal positions. Because of that, some tests are failing.

Desired behavior after this commit is merged:

Said tests are not failing anymore.

task-4369658

closes odoo/enterprise#75844

X-original-commit: 0754d878003130c84481ef969552abbc5772f19d
Related: odoo/odoo#191048
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>
Signed-off-by: Thomas Becquevort (thbe) <thbe@odoo.com>