[IMP] web_gantt,web_map,*: add falsy_value_label on fields
*helpdesk, project_enterprise

This commit is the counterpart of odoo/odoo#190719 which adds an
attribute on field allowing to specify, on a field definition, the
value to display when that field is not set.

We use that attribute in the map and gantt views to display its
value as group name, for the false group.

Moreover, we define it in helpdesk on two fields, which allows to
get rid of several custom views (js_class).

Task~4410227

closes odoo/enterprise#75678

Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>