[IMP] web_gantt: prevent cell clicked if popover opened
When a gantt popover is opened, if one clicks on a cell, a dialog is opened.
This is in general quite anoying since one often just want to close the popover.
Here we solve that problem by using an (already existing) system that allows to
prevent the handling of the click event for a short period of time.

closes odoo/enterprise#74706

Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>