[FIX] website_sale_renting: greyed out add to cart button
Issue
-----
When a customer picks a date range for a rental product that is not valid, the "Add to Cart" button is greyed out and stays greyed out regardless of what the customer does next, even if they change the rental period to one that is available.

Steps
-----
1. Create a new rental product and add at least one pricing in the "Rental prices" tab.
2. Go to the product page on the website and try to pick a date that is not valid (e.g., a range in the past). This should grey out the "Add to Cart" button and display an error message in red.
3. Change the date range to one that is valid. Issue: The "Add to Cart" button stays greyed out.

Cause
-----
In `_verifyValidRentingPeriod` and `_verifyValidPeriod`, when a rental period is not valid, the product form is disabled, and the `css_not_available` class is added to the form. However, this form is never re-enabled, meaning the `css_not_available` class is never removed, and the "Add to Cart" button stays greyed out.

Solution
-----
Call the `_toggleDisable` method every time the rental period is checked to ensure the form is always in the correct state.

opw-4197234

closes odoo/enterprise#75911

X-original-commit: b4d7a015374b31748fe9101b4efb42fe8600e7e9
Signed-off-by: Lionel Piraux (lipi) <lipi@odoo.com>
Signed-off-by: Valentin Chevalier <vcr@odoo.com>