[IMP] knowledge: replace File by EmbeddedFile plugin
Due to changes in the /file command, in html field the result of such
command is a static blue box with a link to a file attachment.

In order to preserve the original behavior, the EmbeddedFile plugin is
available, which modifies the /file command so that it renders the
uploaded file as an embedded component.

This commit makes sure that in the Knowledge app the File plugin (which
produces the static file box) is replaced by the EmbeddedFile plugin
(which produces the embedded component).

task-3522395

X-original-commit: 7e4d1b96876f397640a23c4b9d87cc0364ad60d7
Part-of: odoo/enterprise#77530
Related: odoo/odoo#194609
Signed-off-by: David Monjoie (dmo) <dmo@odoo.com>