[IMP] account_reports: make extra option filter getters smarter
This is kind of bad that each time we add a new filter in the Extra
Option section, we also have to manually override
`hasExtraOptionsFilter` to add that new filter.

With this commit, some intelligence is given to the extra options
filters, by documenting them in the new getter `filterExtraOptionsData`.
So now, the dev that wants to add a new filter will just have to extends
this getter.
He will not need to extend `hasExtraOptionsFilter` or to create the
component manually, this is now done automatically.

task-3676838

closes odoo/enterprise#75878

Signed-off-by: Hugo Poncelet (hupo) <hupo@odoo.com>