[IMP] documents{_spreadsheet}: move actions in a new component for reuse
We move the document actions buttons (share, embedded, ...) in a separated
component so that they can be used elsewhere than in the control panel. We also
clean a bit the DocumentsControlPanel, removing unused functions/variables.

Actually, this commit prepares the addition of the document actions in the
document preview.

Task-4306398

Part-of: odoo/enterprise#74799
Signed-off-by: Stéphane Debauche (std) <std@odoo.com>