[REM] web_gantt: do not use attribute offset
The attribute offset is never used in the gantt view archs and seems less useful
since https://github.com/odoo/enterprise/pull/55618 has been merged. Here we
stop considering that attribute in GanttModel so that we can remove the related
code. That helps for the next commit.

Part-of: odoo/enterprise#74362
Signed-off-by: Bruno Boi (boi) <boi@odoo.com>