[FIX] account_accountant_batch_payment: reconcile partial payment
Steps to reproduce:
- Install accountant
- Create an invoice for $1000
- Register a payment of $100 for the invoice
- Register a second payment of $200 for the invoice
- Create a batch payment with both payments
- Validate the batch payment
- Create a bank statement with an amount equal to the batch payment
(i.e. $300)
- Reconcile the statement with the batch payment

Issue 1:
A UserError is raised stating that the moves are not balanced.
One amount is equal to the sum of the 2 payments (i.e. $300),
but the other one is equal to the amount of the payment term line
of the invoice (i.e. $1000).

Cause 1:
When reconciling a batch payment, only "amount_currency" field
from the invoice lines are used in "_validation_lines_vals".
But in this case, the payments are partial and their sum is not
equal to the amount of of the invoice lines.

Solution 1:
Use the amount from each payment of the batch for the reconciliation
instead of the amount from the account move lines.

Issue 2:
A similar issue happens if the batch payment contains cancelled or
rejected payments.

Solution 2:
Exclude payments with an invalid state from the reconciliation process.

opw-4304772

closes odoo/enterprise#76731

X-original-commit: 01ed19fde39ee929973af97c666367afd5ffe093
Signed-off-by: Ricardo Gomes Rodrigues (rigr) <rigr@odoo.com>
Signed-off-by: Anh Thao Pham (pta) <pta@odoo.com>