[FIX] pos_preparation_display: Prevent Cancel When Payment
This commit is a forward port of 5108b400d1f7c4b7239fbc3f5651cf9b835d385c.

In a POS restaurant order, when reducing the product quantity of an order line, the reduced display line in the preparation display was always the first one created, regardless of which order line was changed.

1. In the POS restaurant, add a product with a quantity of 2 and an internal note. Confirm the order.
2. Add another order with the same product and internal note but with a quantity of 5.
3. In the kitchen display, change the stage of the first display line (quantity 2) from "To Cook" to "Ready."
4. In the POS, reduce the quantity of the second order line (quantity 5) to 4.

Reducing the quantity of the second order line inadvertently reduces the quantity of the first display line (the one in the "Ready" stage).

In `_process_preparation_changes` of `pos_order.py` in the `pos_preparation_display` module, display lines are filtered and sorted based on their creation order. Commit `55f2b24870b664815fe7b926835d07259608ae09` added the `uuid` field, which allows the correct display line to be adjusted when reducing the quantity in the POS order line.

Added a tour to test the following use case fixed by this commit (only in version 18.0 and above):

1. In the POS restaurant, add a product with a quantity of 1 and an internal note. Confirm the order.
2. Add another order with the same product and the same internal note but with a quantity of 2. Reduce the quantity of the second order line from 2 to 1.
3. Proceed to payment. In the preparation display module, the original display orders are canceled, and new ones are created instead.

opw-4354093

closes odoo/enterprise#77299

X-original-commit: 196bbc80a4acac66e9e1daaf7f313d5d49b4e1ef
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>
Signed-off-by: Guillaume Teboul-Tornezy (gute) <gute@odoo.com>