[FIX] helpdesk_timesheet: be able to edit project on timesheet
Before this commit, when the timer is running and linked to a helpdesk
project and a ticket, it was not possible for the current user to
change the project because of the compute triggered for task_id,
helpdesk_ticket_id fields. The reason is because the `helpdesk_ticket_id`
is not removed when the project changed by the user and so the compute
method of project_id field is triggered and will put again the project
of the ticket instead of keeping the changes of the user.

This commit overrides `_onchange_project_id` method in timesheets model
to remove the helpdesk_ticket_id if the project linked to the timesheet
is not the same than the one set on the ticket linked.
It also prevents to recompute the project_id if nothing change inside
the compute of the helpdesk_ticket_id.

Steps to reproduce:
------------------
1. install helpdesk_timesheet module
2. Go to timesheet app
3. Start a timer
4. Select a helpdesk project and a ticket related on the timer header
5. Go to list view of timesheets
6. Try to change the project on the line with the timer running

Expected Behavior
-----------------
The new project selected by the user should be set instead of having the
one of the ticket.

Current Behavior
----------------
The changes made by the user are not taken into account, the project set
is still the initial one (the one selected in step 4)

task-4361124

closes odoo/enterprise#75897

X-original-commit: f4e732c
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>