[REF] knowledge: refactor more options button and panels
Purpose:
--------
This commit refactors the more options button and the panels linked to the
buttons inside it.

The main idea is to use the record object as much as possible instead of relying
on RPC calls to read specific fields.

The more options panel is not a bootstrap dropdown anymore, which leads to the
dropdown being rendered when opened instead of when the topbar is rendered.
This button is now a Component that opens a Dropdown containing DropdownItems.
(OptionsDropdown)

All the original methods linked to the more options panel that were in the
topbar Component are moved inside it.

The display state of each panels has been set as a environment variable inside
the controller. These state variables are reactive Objects linked to owl's
render method.
This replaces the old bus method to be more robust and simpler to understand.

Each panel has also been simplified to not depend on a ref for their rendering.

Task-4273102

Part-of: odoo/enterprise#76261
Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>