[FIX] sale_renting: some list inside the form not taking all space
This commit fixes some ARCH to fixes some visual issue.

Inside a group (grid) there will be always two columns:
* one smaller on the left (max 150px)
* one taking all the remaining space on the right.

When we add only one element with `nolabel="1"` it will take one part of
the grid to take all the space we need to specify it with `colspan="2"`.

Steps to reproduce:
* Go to Sale App
* Go to menu Product
* Select a Product (with Rental)
* Select the `Rental prices` pane
  The list view inside the form not taking all space => BUG

task-4416643

closes odoo/enterprise#75960

Related: odoo/odoo#191313
Signed-off-by: Pierre Paridans (app) <app@odoo.com>