[FIX] web_studio: reportEditor: do not remove spans
Before this commit, spans with no attribute and no special style were unwrapped
by the OdooEditor, meaning the span itself disappeared, leaving its content in the parent element.

This behavior breaks translations, as the whole span is used as a key to match translations

After this commit, those spans are not removed.

opw-3746922
opw-4318712

closes odoo/enterprise#76628

X-original-commit: 4323e23bc20b08da38281c2a00ad58b2defe3b25
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>