[FIX] documents: fix various mobile UI issues
Steps to reproduce:
1. Switch to mobile view.
2. The magnifier button is not working.
3. when adding a member through the share panel, 'people with access' section
shows weirdly.
4. Send and discard button display issue on the share panel.

Technical Reason:
SearchBar is being handled  in the control panel and was not passed the
searchBarToggler in the toggler attribute, and add a condition on the
'searchBarToggler.component' only shows if documents are not selected.

After this Commit:
1. The magnifier button will hide/display the search bar and the magnifier
button will be hidden if we select a document.
2. In the share panel, the 'people with access' section, will show properly.
3. Buttons will be in a proper width.

Task-4419375

closes odoo/enterprise#77182

X-original-commit: 877278fa78360d1bfb71af61282fe347242492c3
Signed-off-by: Florian Charlier (flch) <flch@odoo.com>