[FIX] documents(_project): fix forced False partner on doc upload
Some modules may want to add default values when a `partner_id`
field values is not passed, however this method forced False.

As False is the default value, there is no need to force False,
and it is also easier to remove the default value given by an
override when we do want `False` then the opposite.

Task-4298444

closes odoo/enterprise#75214

Related: odoo/upgrade#6884