[IMP] mail, web, *: `useService()` with auto-`useState()`
Before this commit, when a component uses service "mail.store",
it has to wrap in a `useState()` to make sure any change to store
is observed. This is the behaviour that essentially all
`useService("mail.store")` want. If they omit it, it can fail to
re-render store changed, which is painful to debug.

This commit improves by automatically wrapping `useState()` on
services that are stateful. Stateful services return a reactive.
Using stateful services becomes the same as non-stateful one:
just put `useService(serviceName)`.

closes odoo/enterprise#76594

X-original-commit: 4fa453ece0479aae45c43d5e4bb4a534787f2850
Related: odoo/odoo#192630
Signed-off-by: Sébastien Theys (seb) <seb@odoo.com>
Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>