[FIX] appointment: prevent access error in calendar sync
Before this commit, if a user was an attendee of an event linked to
an appointment type, but did not have access to the appointment (not
part of the staff_user_ids), an access error would occur when trying
to sync their calendar with Google or Microsoft. This error happened
inside the `_get_customer_description` function overridden in the
appointment module.

opw-4402859

closes odoo/enterprise#78439

X-original-commit: a5f910355d9e9326e27b770429bac233061aadcb
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>
Signed-off-by: Pedram Bi Ria (pebr) <pebr@odoo.com>