[FIX] account_accountant_batch_payment: reconciliation with batch of partial payments
An user may receive several partial payments for invoices/bills.
These payments could be grouped in a batch and reconciled with a single
statement line. However this flow is currently not working because the
bank reconciliation fails

Steps to reproduce:
- Create 2 invoice and register a partial payment for each one
- Select the 2 payments create a batch
- Create a new statement with the sum of the payment amounts
- Reconcile the statement with the created batch
Issue: An error will block the validation
```
The move (BNK1/2024/00011) is not balanced.
The total of debits equals $ 510.60 and the total of credits equals $ 100.00.                                                                                                                                                                                                                                                  You might want to specify a default account on journal "Bank" to automatically balance each move.
```

This occurs because the system is taking the amount to reconcile from
the invoice, instead of the payment, so it will try to reconcile the
full amount.

opw-4379673

closes odoo/enterprise#76121

closes odoo/enterprise#76763

Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Andrea Grazioso (agr) <agr@odoo.com>
Co-authored-by: william-andre <wan@odoo.com>