[IMP] website_appointment: add website column on appointment page view
Previously, in the "Appointment Pages" view (Menu bar > Site >
Appointments), there was no way to identify which website a page
belonged to.

This commit adds the "Website" column to the list view, allowing users
to clearly see the associated website for each appointment page.

This change makes it easier for users to work with appointment pages
when they manage multiple websites.

task-4357947

closes odoo/enterprise#76197

Related: odoo/odoo#188992
Signed-off-by: Serge Bayet (seba) <seba@odoo.com>