[IMP] l10n_fr_fec_import: Added the option to import big FEC files
This commit changes the import mechanism of FEC files to support importing big files.
The FEC files are first processed and chunked at the client side.
They are then sent one chunk at a time to the server side.
The chunking ensures that no memory or time limit at the server side are ever exceeded.

Because some of the processing is done in JS side and the rest in Py, the tests cover both sides.
A tour was also added to test the integration between the two sides.
The FEC test files in `fec_test_files` folder were named with the following convension:
fec_[new-line character]_[fields delimiter]_[encoding]_[extra details about the file if any]
The files with .txt or .csv extension are regular FEC files.
The files with .json extension are their equivalent parsed files by the JS code in FecParser.
The JSON files are used for output validation in JS tests, and they are used as input files for Py tests.

task-3597066

closes odoo/enterprise#76216

Related: odoo/upgrade#7235
Signed-off-by: William André (wan) <wan@odoo.com>