[FIX] helpdesk: sla report status clean up the graph view
This commit's purpose is to clean up the graph view of the sla report analysis model. Some fields were added on the view by a generic improvement task to handle the widget 'float_time' with grap view. In this case, those fields are a nuisance for readability, so we're getting rid of them.

target version 17.0 - master
task - 4351830

closes odoo/enterprise#77338

X-original-commit: 58a9b86e9efd82b2541fcef45591998b49d2a0cc
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>