[REF] l10n_cl_edi: fetch_mail aligned with refactor in odoo
Align the function `_fetch_mail` with the refactoring in the community
repository. We add a savepoint to handle errors properly.

Note: this does not implement the progress notification for cron jobs as
we do not know the global state here, the recommendation is to move the
processing of e-mails to `message_process` as we now handle the case
where there is a lot of new e-mails on an IMAP server.

closes odoo/enterprise#76258

Related: odoo/odoo#191911
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>