[IMP] project: changed budget column headings and update budget progress
Before this commit:
The budget columns included 'Allocated', 'Spent' and
'Remaining.' The budget progress indicator classified budgets as 'on track',
'soon overspent', and 'overspent' based on the threshold values.

After this commit:
The budget columns have been updated to 'Forecast'
'Actual' and 'Difference' for better clarity. The budget progress indicator
has been refined to include a 'budget allocated' status when the budget is just
assigned and not spent on anything, to give a more accurate representation
of the budget state.

Task-4353191

closes odoo/enterprise#76320

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>