[FIX] delivery_fedex_rest: Add person name in the request when partner is company
Before this commit:

Only `companyName` is included in shipping request sent to fedex.

After this commit:

`personName` is added to the request, same as `companyName`.

According to fedex documentation it is not required if the address is `buisness` but we get an error if the
`deliveryAddress` and 'soldTo` addresses are different.

Checked with fedex support and they said it is required to send `personName` field always.

opw-4415536

closes odoo/enterprise#76363

X-original-commit: daee25b8eebc08dd7df8a2bf6b159ccd338ffcc5
Signed-off-by: Tiffany Chang (tic) <tic@odoo.com>
Signed-off-by: Basit Arshad (baar) <baar@odoo.com>