[IMP] knowledge: allow users to add articles to the templates gallery
This commit adds a new button to the Knowledge kebab menu, allowing users
to add articles to the templates gallery. This feature improves usability
by making it easier for users to locate their templates and create copies.

To ease template management, a delete button has been added next to each
article and template in the gallery. The button is displayed only to
users with the appropriate permissions:

- Deleting a template requires administrator privileges.
- Deleting an article from the templates gallery requires "write" access
  to the article.

Task-4247544

closes odoo/enterprise#76306

Signed-off-by: Damien Abeloos (abd) <abd@odoo.com>