[FIX] l10n_pe_edi: withholding amount sign is PEN in PDF
Scenario:

- Install l10n_pe_edi and switch to "PE Company"
- Create a product with "Withhold Percentage" 20%
- Create an invoice with operation type "[1001] Operation Subject to
  Detraction", the created product and USD currency.
- Print the invoice as PDF

=> The "Withholding amount" in the PDF is correctly shown in the company
   currency, but the currency symbol is the one of the invoice ($)

Issue:

Amount is always in the company currency, and currency is the invoice
currency that might be different.

Fix:

currency_id gotten from account.move()._l10n_pe_edi_get_spot is only
used in this location where the amount should always be the company
currency: so we modify it to be the company currency.

opw-4292774
opw-4338728

closes odoo/enterprise#76922

X-original-commit: 9725e5c5f9c829ba6bad7d878bdbc3347ce650ae
Signed-off-by: Claire Bretton (clbr) <clbr@odoo.com>