[FIX] sale_subscription_stock: test's start date issue
Before this commit, the test was relying in "today's" date for
creating the recurring invoices. Since the time period for
generating them was too short, the expected deliveries were
not being generated.

After this commit, we give a fixed date to the start date
of the subscriptions and then generate the recurring invoices
with a time distance of a month, solving the issue.

Issue-from: OPW-4166852
Related PR: odoo/enterprise#73397

closes odoo/enterprise#79263

X-original-commit: c0da9b77279c5e44cc3396328c6e3ab809d9dbe9
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: William Braeckman (wbr) <wbr@odoo.com>
Signed-off-by: Gabriel de Paula Felix (gdpf) <gdpf@odoo.com>