[IMP] (account)(iap)extract: allow manual correction of OCRed bank statements
After a bank statement has been OCRed, we allow the user
to make corrections by simply clicking one of the balance
fields and then clicking the correct amount by clicking
on the boxes overlay of the imported file.

This is a feature that was introduced first for bills, and
is now made available to bank statements as well. A new mixin
has been written to factorize as much as possible between
the two models.

task-id 4260464

closes odoo/enterprise#76437

Related: odoo/upgrade#7010
Related: odoo/upgrade-util#182
Signed-off-by: Ricardo Gomes Rodrigues (rigr) <rigr@odoo.com>