[FIX] iot: IoT Box couldn't download lib/ handlers
In this [PR](https://github.com/odoo/enterprise/pull/75443), we added conditions not to
download already up to date handlers.
We introduced a condition to only download Windows handlers on Windows and Linux ones on Linux:
as some are not plateform-specific, we were not sending them to the IoT Box.

The condition is now adapted to download both plateform-specific and generic handlers to IoT Boxes.

closes odoo/enterprise#76386

Signed-off-by: Yaroslav Soroko (yaso) <yaso@odoo.com>